MANUAL: Improve documentation on using python code in the config
Hi, the following patch updates the documentation.  It was a bit out of
context since Sebastian originally just copied it from my webpage.
Also, the code for the helper functions is available there under
"Download".  I did not include it because it would leave larger chunks
of python code in the manual which seems weird.

Signed-off-by: Thomas Kahle <tomka@gentoo.org>
Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>